requirements.txt
alembic==1.4.2
aniso8601==8.0.0
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
decorator==4.4.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.3
flask-restplus==0.13.0
flask-restx==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.8.0
idna==2.9
importlib-metadata==1.6.0
ipython==7.14.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
jsonschema==3.2.0
lightgbm==2.3.1
Mako==1.1.2
MarkupSafe==1.1.1
numpy==1.18.4
pandas==1.0.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pyrsistent==0.16.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
redis==3.5.2
requests==2.23.0
rq==1.4.0
scikit-learn==0.23.1
scipy==1.4.1
six==1.14.0
SQLAlchemy==1.3.17
threadpoolctl==2.0.0
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
xgboost==0.90
zipp==3.1.0
